+++ R/ru_sampling.R
@@ -140,7 +140,11 @@
 #'     \item{pa}{A numeric scalar.  An estimate of the probability of
 #'       acceptance.}
 #'     \item{d}{A numeric scalar.  The dimension of \code{logf}.}
-#'     \item{logf}{A function. \code{logf} function supplied by the user.}
+#'     \item{logf}{A function. \code{logf} supplied by the user, but
+#'       with f scaled by the maximum of the target density used in the
+#'       ratio-of-uniforms method (i.e. \code{logf_rho}), to avoid numerical
+#'       problems in contouring fin \code{\link{plot.ru}} when
+#'       \code{d = 2}.}
 #'     \item{logf_rho}{A function. The target function actually used in the
 #'       ratio-of-uniforms algorithm.}
 #'     \item{sim_vals_rho}{An \code{n} by \code{d} matrix of values simulated
@@ -645,6 +649,7 @@ ru <- function(logf, ..., n = 1, d = 1, init = NULL,
     print(res$box)
   }
   res$d <- d
+  # Return logf - hscale to avoid over/under-flow in plot.ru() when d = 2.
   logf_scaled <- function(theta, ...) {
     return(logf(theta, ...) - hscale)
   }