+++ R/ru_sampling_rcpp.R
@@ -149,7 +149,11 @@
 #'     \item{pa}{A numeric scalar.  An estimate of the probability of
 #'       acceptance.}
 #'     \item{d}{A numeric scalar.  The dimension of \code{logf}.}
-#'     \item{logf}{A function. \code{logf} function supplied by the user.}
+#'     \item{logf}{A function. \code{logf} supplied by the user, but
+#'       with f scaled by the maximum of the target density used in the
+#'       ratio-of-uniforms method (i.e. \code{logf_rho}), to avoid numerical
+#'       problems in contouring f in \code{\link{plot.ru}} when
+#'       \code{d = 2}.}
 #'     \item{logf_rho}{A function. The target function actually used in the
 #'       ratio-of-uniforms algorithm.}
 #'     \item{sim_vals_rho}{An \code{n} by \code{d} matrix of values simulated
@@ -723,7 +727,11 @@ ru_rcpp <- function(logf, ..., n = 1, d = 1, init = NULL,
     print(res$box)
   }
   res$d <- d
-  res$logf <- cpp_logf
+  # Add hscale to pars (and hence res$logf_args) and return cpp_logf_scaled
+  # (which is cpp_logf - hscale) rather than cpp_logf.
+  # This is to avoid over/under-flow in plot.ru() when d = 2.
+  pars$hscale <- logf_args$hscale
+  res$logf <- cpp_logf_scaled
   res$logf_args <- list(logf = logf, pars = pars)
   res$logf_rho <- logf_fun
   res$logf_rho_args <- logf_args