+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Bug fixes and minor improvements
 
-* In `ru()` the constant `hscale` used to shift `logf` (and hence scale the target density f) is now included in function `logf` in the returned object.  This helps to avoid over/under-flow when `logf` is calculated latterly, e.g. in `plot()`.  This hasn't been done for `ru_rcpp()` (yet).
+* In `ru()` and `ru_rcpp()` the constant `hscale` is used to shift `logf` (and hence scale the target density f) in function `logf` in the returned object.  This helps to avoid over/under-flow when contouring f in `plot.ru` when `d = 2`.
 
 * The `var_names` argument to `ru_rcpp` didn't work.  This has been corrected.